+++ frontend/src/components/settings/index.ts
@@ -1,0 +1,12 @@
+export { AccountSection } from './AccountSection';
+export { LicenseSection } from './LicenseSection';
+export { UsersSection } from './UsersSection';
+export { SystemSection } from './SystemSection';
+export { NotificationsSection } from './NotificationsSection';
+export { WebhooksSection } from './WebhooksSection';
+export { DeveloperSection } from './DeveloperSection';
+export { AppStoreSection } from './AppStoreSection';
+export { SupportSection } from './SupportSection';
+export { AboutSection } from './AboutSection';
+export { DEFAULT_SETTINGS } from './types';
+export type { PatchableSettings, SectionId, Agent } from './types';
